refactor: add fk conditional

+++ src/data/services/RoleGrade/DeleteRoleGrade.service.js
@@ -1,8 +1,12 @@
-import { RoleGradeRepository } from '../../database/repositories';
+import {
+  RoleGradeRepository,
+  ProfessionalRepository,
+} from '../../database/repositories';
 
 export class DeleteRoleGradeService {
   async execute({ id_role_grade }) {
     const repository = new RoleGradeRepository();
+    const professionalRepository = new ProfessionalRepository();
 
     const verifyRoleExists = await repository.findRoleGradeById({
       id_role_grade,
@@ -12,6 +16,19 @@ export class DeleteRoleGradeService {
       return {
         error: `Não existe um Custo HH com este ID -> ${id_role_grade}.`,
       };
+
+    const verifyFkFromProfessional = await professionalRepository.verifyRelationRoleGrade(
+      {
+        id_role_grade,
+      }
+    );
+
+    if (verifyFkFromProfessional.length > 0) {
+      return {
+        error:
+          'Não foi possível excluir o Custo HH pois existem Colaboradores associados.',
+      };
+    }
 
     await repository.deleteRoleGrade({
       id_role_grade,
