+++ src/data/services/Sector/DeleteSector.service.js
@@ -1,8 +1,12 @@
-import { SectorRepository } from '../../database/repositories';
+import {
+  SectorRepository,
+  ProfessionalRepository,
+} from '../../database/repositories';
 
 export class DeleteSectorService {
   async execute({ id_sector }) {
     const repository = new SectorRepository();
+    const professionalRepository = new ProfessionalRepository();
 
     const verifySectorExists = await repository.findSectorById({
       id_sector,
@@ -10,6 +14,19 @@ export class DeleteSectorService {
 
     if (!verifySectorExists)
       return { error: `Não existe um Setor com este ID -> ${id_sector}.` };
+
+    const verifyFkFromProfessional = await professionalRepository.verifyRelationSector(
+      {
+        id_sector,
+      }
+    );
+
+    if (verifyFkFromProfessional.length > 0) {
+      return {
+        error:
+          'Não foi possível excluir o Setor pois existem Colaboradores associados.',
+      };
+    }
 
     await repository.deleteSector({
       id_sector,
